delivery-backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==3.2.7
django-allauth==0.48.0
django-cors-headers==3.11.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-taggit==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
oauthlib==3.2.0
phonenumbers==8.12.42
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.8
zipp==3.7.0
